refactor(kubernetes-ingestor): improve plugin structure and organization
BREAKING CHANGE: File structure reorganization - imports may need updating

- Split EntityProvider.ts into separate files for better maintainability
  - Created XRDTemplateEntityProvider.ts for XRD/CRD template generation
  - Created KubernetesEntityProvider.ts for K8s resource ingestion
  - Extracted BackstageLink interface to its own file

- Improved folder organization:
  - interfaces/ - Contains type definitions and interfaces
  - entity-providers/ - Contains entity provider implementations
  - data-providers/ - Contains data fetching providers
  - services/ - Contains core services

- Added comprehensive architecture documentation with:
  - Flow diagrams showing component interactions
  - Class dependency diagrams
  - Data flow sequence diagrams
  - Configuration examples
  - Troubleshooting guide

- Maintained backward compatibility through barrel exports
- Fixed TypeScript isolatedModules issue with interface exports
- All tests and builds passing successfully

This refactoring follows single responsibility principle with one class
per file, making the codebase more maintainable and easier to navigate.
