+++ plugins/kubernetes-ingestor/src/entity-providers/index.ts
@@ -1,0 +1,2 @@
+export { XRDTemplateEntityProvider } from './XRDTemplateEntityProvider';
+export { KubernetesEntityProvider } from './KubernetesEntityProvider';
