+++ plugins/kubernetes-ingestor/src/interfaces/BackstageLink.ts
@@ -1,0 +1,6 @@
+export interface BackstageLink {
+  url: string;
+  title: string;
+  icon: string;
+  [key: string]: string;
+}
