+++ plugins/kubernetes-ingestor/src/providers/index.ts
@@ -1,4 +1,4 @@
-export { KubernetesEntityProvider, XRDTemplateEntityProvider } from './EntityProvider';
+export { KubernetesEntityProvider, XRDTemplateEntityProvider } from '../entity-providers';
 export { CRDDataProvider } from './CRDDataProvider';
 export { XRDDataProvider } from './XRDDataProvider';
 export { KubernetesDataProvider } from './KubernetesDataProvider';
